requirements.txt
# Web framework
flask>=3.0.0
flask-cors>=4.0.0

# Anthropic Claude API
anthropic>=0.18.0

# Google Cloud TTS
google-cloud-texttospeech>=2.14.0

# Database
# (SQLite is built into Python)

# Utilities
python-dotenv>=1.0.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
